Fix #41- applying spring data jpa in member repository
