+++ jpashop/src/main/java/jpabook/jpashop/service/OrderService.java
@@ -12,7 +12,7 @@ import jpabook.jpashop.domain.OrderItem;
 import jpabook.jpashop.domain.OrderSearch;
 import jpabook.jpashop.domain.item.Item;
 import jpabook.jpashop.repository.ItemRepository;
-import jpabook.jpashop.repository.MemberRepository;
+import jpabook.jpashop.repository.MemberRepositoryOld;
 import jpabook.jpashop.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
 
@@ -22,7 +22,7 @@ import lombok.RequiredArgsConstructor;
 public class OrderService {
 
 	private final OrderRepository orderRepository;
-	private final MemberRepository memberRepository;
+	private final MemberRepositoryOld memberRepositoryOld;
 	private final ItemRepository itemRepository;
 
 	/**
@@ -31,7 +31,7 @@ public class OrderService {
 	@Transactional
 	public Long order(Long memberId, Long itemId, int count) {
 		// 엔티티 조회
-		Member member = memberRepository.findOne(memberId);
+		Member member = memberRepositoryOld.findOne(memberId);
 		Item item = itemRepository.findOne(itemId);
 
 		// 배송 정보 생성
